Add DOMupdates file and methods
